refactor(shared): promote gameweek-selector out of the teams domain
P2.5. Four domains — admin, leagues, players and transfers — imported
GameweekSelector from teams/components/, which Rule 2 allowlisted as
reaching inside another domain. MAY_REACH_INSIDE 21 -> 17.

It qualified for _shared/components/ without argument: its entire import
list was react-router, _shared/lib/fpl/fpl-types and its own stylesheet.
Zero domain imports, so this is a move and five path rewrites with no
judgement call about what the component means. A gameweek picker is not
teams-domain UI; it only lived there because teams needed it first.

The fifth importer is teams/components/team-view.tsx, which owned it.

No cycle dissolved: teams <-> leagues survives on
leagues/server/team-of-the-week.server -> teams/types/team-types.

The larger half of P2.5 — players/components/player, 8 importers — is
deliberately not in here. It imports EnhancedPlayerData and, via
DisplayablePlayer, RosterPlayer, so moving it to _shared today would add
new _shared -> domain edges and push that count from 3 back to ~5. It is
blocked on P2.1b; the reasoning and the expose-vs-promote decision are
written up in the backlog.

Tests 346 pass, ratchet counts all unchanged (correct for a pure move),
yarn build clean.
