+++ draft/app/_shared/components/gameweek-selector.module.css
@@ -1,4 +1,4 @@
-/* Location: app/teams/components/gameweek-selector.module.css */
+/* Location: app/_shared/components/gameweek-selector.module.css */
 
 .gameweekSelector {
     display: flex;
