+++ draft/app/leagues/league-standings.tsx
@@ -1,12 +1,12 @@
 /* Location: app/leagues/league-standings.tsx */
 
 import { useLoaderData, useNavigate, useSearchParams } from 'react-router';
+import { GameweekSelector } from '../_shared/components/gameweek-selector';
 import { PageHeader } from '../_shared/components/page-header';
 import { SelectDivision } from '../_shared/components/select-division';
 import { TimeTravelBanner } from '../_shared/components/time-travel-banner';
 import { UserSelectionProvider } from '../_shared/features/user-selection/user-selection-provider';
 import type { DivisionId } from '../_shared/types/league-types';
-import { GameweekSelector } from '../teams/components/gameweek-selector';
 import { PositionPointsTable } from './components/position-points-table';
 import type { EnhancedLeagueStandingsLoaderData } from './types/league-standings-types';
